CitableObject.jl Documentation

CitableObject.jl Documentation

Types

Functions

Index

Types

CitableObject.Cite2Urn — Type

Implementation of the Cite Object URN. See http://cite-architecture.org/ctsurn/.

Functions

CitableObject.namespace — Function

namespace(u)

Extract CITE2 namespace from a CtsUrn.

Examples

julia>
namespace(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.collectioncomponent — Function

collectioncomponent(u)

Extract CITE2 work component from a CtsUrn.

Examples

julia>
collectioncomponent(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.objectcomponent — Function

objectcomponent(u)

Extract CITE2 object component from a CtsUrn.

Examples

julia>
objectcomponent(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.isrange — Function

isrange(obj)

True if obj is a range expression.

isrange(urn)

True if the object component of urn is a range expression.

CitableObject.rangebegin — Function

rangebegin(s)

Extract first part from a range expression.

rangebegin(u)

Extract first range part from a URN's object identifier.

CitableObject.rangeend — Function

rangeend(s)

Extract second part from a range expression.

rangeend(urn)

Extract second range part from a URN's object identifier.

CitableObject.hassubref — Function

hassubref(obj)

True if String has a subreference.

hassubref(urn)

True if urn has a subreference.

CitableObject.subref — Function

subref(s)

Extract subreference from a string.

subref(u)

Extract subreference from a URN.

CitableObject.dropsubref — Function

dropsubref(s)

Remove any subreference substrings in a string. If s is a range expression, separately drop subreference from each of range-begin and range-end parts.

dropsubref(u)

Remove any subreferences from a CtsUrn.

CitableBase.dropversion — Function

Catch-all function for dropversion method.

dropversion(u)

Create a new Cite2Urn by dropping the version part of a given Cite2Urn.

dropversion(u)

CitableBase.addversion — Function

Catch-all function for addversion method.

addversion(u, s)

Create a new Cite2Urn by setting the version part of a given Cite2Urn to a given value.

addversion(u, versid)

Index

CitableObject.Cite2Urn

CitableBase.addversion

CitableBase.dropversion

CitableObject.collectioncomponent

CitableObject.dropsubref

CitableObject.hassubref

CitableObject.isrange

CitableObject.namespace

CitableObject.objectcomponent

CitableObject.rangebegin

CitableObject.rangeend

CitableObject.subref