API Documentation

Types

CitableObject.Cite2Urn — Type

Implementation of the Cite Object URN. See http://cite-architecture.org/cite2urn/.

CitableObject.CitableByCite2Urn — Type

A value of CitableBase.CitableTrait that can be used for content citable by CtsUrn.

CitableObject.Cite2UrnComparable — Type

Value for URN comparison trait on CtsUrnDocStringExtensions

Functions

Manipulating Cite2Urns:

CitableObject.namespace — Function

namespace(u)

Extract CITE2 namespace from a CtsUrn.

Examples

julia>
namespace(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.collectioncomponent — Function

collectioncomponent(u)

Extract CITE2 work component from a CtsUrn.

Examples

julia>
collectioncomponent(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.objectcomponent — Function

objectcomponent(u)

Extract CITE2 object component from a CtsUrn.

Examples

julia>
objectcomponent(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.isrange — Function

isrange(obj)

True if obj is a range expression.

isrange(urn)

True if the object component of urn is a range expression.

CitableObject.rangebegin — Function

rangebegin(s)

Extract first part from a range expression.

rangebegin(u)

Extract first range part from a URN's object identifier.

CitableObject.rangeend — Function

rangeend(s)

Extract second part from a range expression.

rangeend(urn)

Extract second range part from a URN's object identifier.

CitableBase.supportssubref — Function

True for URN types that support subreferences.

supportssubref(u)

Cite2Urns support subreferences.

supportssubref(u)

CitableBase.hassubref — Function

True for URN types that support subreferences.

hassubref(u)

hassubref(obj)

True if String has a subreference.

hassubref(urn)

True if urn has a subreference.

CitableBase.subref — Function

Urn subtypes should implement dropversion(urn::U)::U.

subref(u)

subref(s)

Extract subreference from a string.

subref(u)

Extract subreference from a URN.

CitableBase.dropsubref — Function

Urn subtypes should implement dropversion(urn::U)::U.

dropsubref(u)

dropsubref(s)

Remove any subreference substrings in a string. If s is a range expression, separately drop subreference from each of range-begin and range-end parts.

dropsubref(u)

Remove any subreferences from a CtsUrn.

CitableBase.supportsversion — Function

True for URN types that support versioning.

supportsversion(u)

Cite2Urns support version ids.

supportsversion(u)

CitableBase.dropversion — Function

Urn subtypes should implement dropversion(urn::U)::U.

dropversion(u)

Create a new Cite2Urn by dropping the version part of a given Cite2Urn.

dropversion(u)

CitableBase.addversion — Function

Urn subtypes should implement addversion(urn::U, versionid)::U.

addversion(u, versioninfo)

Create a new Cite2Urn by setting the version part of a given Cite2Urn to a given value.

addversion(u, versid)

CitableBase.versionid — Function

Urn subtypes should implement addversion(urn::U, versionid)::U.

versionid(u)

versionid(u)

Extract CITE2 collection identifer from a Cite2Urn.

Examples

julia>
urn = Cite2Urn("urn:cite2:hmt:msA.v1:12r")
collectionid(urn)

CitableObject.dropproperty — Function

Create a new Cite2Urn by dropping the version part of a given Cite2Urn.

dropproperty(u)

CitableObject.addproperty — Function

Create a new Cite2Urn by setting the version part of a given Cite2Urn to a given value.

addproperty(u, propid)

Comparing a pair of Cite2Urns:

CitableBase.urncontains — Function

URN-comparable objects must implement urncontains.

urncontains(x, y)

urncontains(u1, u2)

True if urn1 contains or is equal to urn2 accroding to URN logic.

```

CitableBase.urnequals — Function

For subtypes of Urn, implement urnequals concretely.

urnequals(u1, u2)

URN-comparable objects must implement urnequals.

urnequals(urn1, urn2)

Override urnequals function for Cite2Urns.

CitableBase.urnsimilar — Function

URN-comparable objects must implement urnsimilar.

urnsimilar(urn1, urn2)

Override similar function to be true if urn1 and urn1 are similar by URN logic.

CexUtils submodule

Finding data by URN value:

CitableObject.CexUtils.properties — Function

Find all property definitions in the CEX string s contained by u.

properties(s, u)

Find all property definitions in blks contained by u.

properties(blks, u)

Dispatch properties on T.

properties(cexsrc, u, reader)

Find all property definitions in CEX file filesrc contained by u.

properties(filesrc, u, freader)

Find all property definitions in CEX at URL url contained by u.

properties(url, u, ureader)

Find all property definitions in CEX at URL url contained by u.

properties(s, u, sreader)

CitableObject.CexUtils.collectiondata — Function

Find all data lines in blkscontained byu`.

collectiondata(blks, u; delimiter)

Find all data lines in the CEX string s contained by u.

collectiondata(s, u; delimiter)

Dispatch collectiondata on T.

Find all data lines in CEX file filesrc contained by u.

collectiondata(filesrc, u, freader; delimiter)

Find all data lines in CEX at URL url contained by u.

collectiondata(url, u, ureader; delimiter)

Find all property definitions in CEX at URL url contained by u.

collectiondata(s, u, sreader; delimiter)

CitableObject.CexUtils.implementations — Function

Find Cite2Urns for all collections in s implementing datamodel.

implementations(s, datamodel; delimiter)

Find Cite2Urns for all collections in blks implementing datamodel.

implementations(blks, datamodel; delimiter)

Dispatch implementations on T.

Find Cite2Urns for all collections in CEX file filesrc implementing datamodel.

implementations(filesrc, datamodel, freader; delimiter)

Find Cite2Urns for all collections in CEX file filesrc implementing datamodel.

implementations(url, datamodel, freader; delimiter)

Find Cite2Urns for all collections in CEX file filesrc implementing datamodel.

implementations(s, datamodel, freader; delimiter)

CitableObject.CexUtils.data_for_model — Function

Find all data lines in the CEX string s for collections implementing datamodel.

data_for_model(s, datamodel; delimiter)

Find all data lines in blks for collections implementing datamodel.

data_for_model(blks, datamodel; delimiter)

Dispatch data_for_model on T.

Find all data lines in file filesrc for collections implementing datamodel.

data_for_model(filesrc, datamodel, freader; delimiter)

Find all data lines in content of url implementing datamodel.

data_for_model(url, datamodel, freader; delimiter)

Find all data lines in s implementing datamodel.

data_for_model(s, datamodel, freader; delimiter)

CitableObject.CexUtils.relations_for_model — Function

Find relations data for all relation sets in s implementing data model dm.

relations_for_model(s, dm)

Find relations data for all relation sets in blks implementing data model dm.

relations_for_model(blks, dm)

Dispatch relations on T.

Find relations data for all relation sets in CEX file filesrc implementing data model dm.

relations_for_model(filesrc, dm, freader)

Find relations data for all relation sets from CEX at URL url implementing data model dm.

relations_for_model(url, dm, freader)

Find relations data for all relation sets in s implementing data model dm.

relations_for_model(s, dm, freader)

Finding human-readable labels by URN value:

CitableObject.CexUtils.cataloglabel — Function

Find catalog label, if any, for u in CEX source s. If no catalog found, compose machine-generated label.

cataloglabel(s, u; delimiter)

Find catalog label, if any, for u in CEX source s. If no catalog found, compose machine-generated label.

cataloglabel(blks, u; delimiter)

Dispatch properties on T.

Find catalog label, if any, in CEX file filesrc contained by u.

cataloglabel(filesrc, u, freader)

Find catalog label, if any, in CEX at URL url contained by u.

cataloglabel(url, u, freader)

Find catalog label, if any, in CEX at URL url contained by u.

cataloglabel(s, u, sreader)