API Documentation

Types

CitableObject.Cite2Urn — Type

Implementation of the Cite Object URN. See http://cite-architecture.org/cite2urn/.

Functions

Manipulating Cite2Urns:

CitableObject.namespace — Function

namespace(u)

Extract CITE2 namespace from a CtsUrn.

Examples

julia>
namespace(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.collectioncomponent — Function

collectioncomponent(u)

Extract CITE2 work component from a CtsUrn.

Examples

julia>
collectioncomponent(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.objectcomponent — Function

objectcomponent(u)

Extract CITE2 object component from a CtsUrn.

Examples

julia>
objectcomponent(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.isrange — Function

isrange(obj)

True if obj is a range expression.

isrange(urn)

True if the object component of urn is a range expression.

CitableObject.rangebegin — Function

rangebegin(s)

Extract first part from a range expression.

rangebegin(u)

Extract first range part from a URN's object identifier.

CitableObject.rangeend — Function

rangeend(s)

Extract second part from a range expression.

rangeend(urn)

Extract second range part from a URN's object identifier.

CitableObject.hassubref — Function

hassubref(obj)

True if String has a subreference.

hassubref(urn)

True if urn has a subreference.

CitableObject.subref — Function

subref(s)

Extract subreference from a string.

subref(u)

Extract subreference from a URN.

CitableObject.dropsubref — Function

dropsubref(s)

Remove any subreference substrings in a string. If s is a range expression, separately drop subreference from each of range-begin and range-end parts.

dropsubref(u)

Remove any subreferences from a CtsUrn.

CitableBase.dropversion — Function

Urn subtypes should implement dropversion(urn::U)::U.

CitableBase.addversion — Function

Urn subtypes should implement addversion(urn::U, versionid)::U.

CitableObject.dropproperty — Function

Create a new Cite2Urn by dropping the version part of a given Cite2Urn.

dropproperty(u)

CitableObject.addproperty — Function

Create a new Cite2Urn by setting the version part of a given Cite2Urn to a given value.

addproperty(u, propid)

Comparing a pair of Cite2Urns:

CitableBase.urncontains — Function

URN-comparable objects must implement urncontains.

urncontains(x)

urncontains(u1, u2)

True if urn1 contains or is equal to urn2 accroding to URN logic.

```

CitableBase.urnequals — Function

URN-comparable objects must implement urnequals.

urnequals(urn1, urn2)

Override urnequals function for Cite2Urns.

CitableBase.urnsimilar — Function

URN-comparable objects must implement urnsimilar.

urnsimilar(urn1, urn2)

Override similar function to be true if urn1 and urn1 are similar by URN logic.