API Documentation

Types

CitableObject.Cite2Urn — Type

Implementation of the Cite Object URN. See http://cite-architecture.org/cite2urn/.

CitableObject.CitableByCite2Urn — Type

A value of CitableBase.CitableTrait that can be used for content citable by CtsUrn.

CitableObject.Cite2UrnComparable — Type

Value for URN comparison trait on CtsUrnDocStringExtensions

Functions

Manipulating Cite2Urns:

CitableObject.namespace — Function

namespace(u)

Extract CITE2 namespace from a CtsUrn.

Examples

julia>
namespace(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.collectioncomponent — Function

collectioncomponent(u)

Extract CITE2 work component from a CtsUrn.

Examples

julia>
collectioncomponent(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.objectcomponent — Function

objectcomponent(u)

Extract CITE2 object component from a CtsUrn.

Examples

julia>
objectcomponent(Cite2Urn("urn:cite2:hmt:msA.v1:12r"))

CitableObject.isrange — Function

isrange(obj)

True if obj is a range expression.

isrange(urn)

True if the object component of urn is a range expression.

CitableObject.rangebegin — Function

rangebegin(s)

Extract first part from a range expression.

rangebegin(u)

Extract first range part from a URN's object identifier.

CitableObject.rangeend — Function

rangeend(s)

Extract second part from a range expression.

rangeend(urn)

Extract second range part from a URN's object identifier.

CitableBase.supportssubref — Function

True for URN types that support subreferences.

supportssubref(u)

Cite2Urns support subreferences.

supportssubref(u)

CitableBase.hassubref — Function

True for URN types that support subreferences.

hassubref(u)

hassubref(obj)

True if String has a subreference.

hassubref(urn)

True if urn has a subreference.

CitableBase.subref — Function

Urn subtypes should implement dropversion(urn::U)::U.

subref(u)

subref(s)

Extract subreference from a string.

subref(u)

Extract subreference from a URN.

CitableBase.dropsubref — Function

Urn subtypes should implement dropversion(urn::U)::U.

dropsubref(u)

dropsubref(s)

Remove any subreference substrings in a string. If s is a range expression, separately drop subreference from each of range-begin and range-end parts.

dropsubref(u)

Remove any subreferences from a CtsUrn.

CitableBase.supportsversion — Function

True for URN types that support versioning.

supportsversion(u)

Cite2Urns support version ids.

supportsversion(u)

CitableBase.dropversion — Function

Urn subtypes should implement dropversion(urn::U)::U.

dropversion(u)

Create a new Cite2Urn by dropping the version part of a given Cite2Urn.

dropversion(u)

CitableBase.addversion — Function

Urn subtypes should implement addversion(urn::U, versionid)::U.

addversion(u, versioninfo)

Create a new Cite2Urn by setting the version part of a given Cite2Urn to a given value.

addversion(u, versid)

CitableBase.versionid — Function

Urn subtypes should implement addversion(urn::U, versionid)::U.

versionid(u)

versionid(u)

Extract CITE2 collection identifer from a Cite2Urn.

Examples

julia>
urn = Cite2Urn("urn:cite2:hmt:msA.v1:12r")
collectionid(urn)

CitableObject.dropproperty — Function

Create a new Cite2Urn by dropping the version part of a given Cite2Urn.

dropproperty(u)

CitableObject.addproperty — Function

Create a new Cite2Urn by setting the version part of a given Cite2Urn to a given value.

addproperty(u, propid)

Comparing a pair of Cite2Urns:

CitableBase.urncontains — Function

URN-comparable objects must implement urncontains.

urncontains(x, y)

urncontains(u1, u2)

True if urn1 contains or is equal to urn2 accroding to URN logic.

```

CitableBase.urnequals — Function

For subtypes of Urn, implement urnequals concretely.

urnequals(u1, u2)

URN-comparable objects must implement urnequals.

urnequals(urn1, urn2)

Override urnequals function for Cite2Urns.

CitableBase.urnsimilar — Function

URN-comparable objects must implement urnsimilar.

urnsimilar(urn1, urn2)

Override similar function to be true if urn1 and urn1 are similar by URN logic.

CexUtils submodule

CitableObject.CexUtils.properties — Function

Find all property definitions in the CEX string s contained by u.

properties(s, u; delimiter)

Find all property definitions in blks contained by u.

properties(blks, u; delimiter)

Missing docstring.

Missing docstring for CitableObject.CexUtils.collectiondata. Check Documenter's build log for details.

CitableObject.CexUtils.relations — Function

Missing docstring.

Missing docstring for CitableObject.CexUtils.collections_for_model. Check Documenter's build log for details.